# combined_swagger.yaml
openapi: 3.0.3
info:
  title: TAAFT Backend API
  description: API for the TAAFT (Tools Analysis and Functionality Testing) application
  version: 1.0.0
  contact:
    name: API Support
    email: support@taaft.com
servers:
- url: http://localhost:8000
  description: Local development server
- url: https://api.taaft.com
  description: Production server
...
paths:
...
  /api/search/tools:
    get:
      tags:
      - Search
      summary: Search tools
      description: Search for tools using Algolia
      operationId: searchTools
      parameters:
      - name: query
        in: query
        required: false
        schema:
          type: string
        description: Search query
      - name: categories
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
        style: form
        explode: true
        description: List of category IDs to filter by
      - name: pricing_types
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
        style: form
        explode: true
        description: List of pricing types to filter by
      - name: min_rating
        in: query
        required: false
        schema:
          type: number
          minimum: 0
          maximum: 5
        description: Minimum rating to filter by
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
        description: Page number (1-based)
      - name: per_page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        description: Number of results per page
      - name: sort_by
        in: query
        required: false
        schema:
          type: string
          enum:
          - relevance
          - newest
          - trending
        description: Sort order
      - name: filters
        in: query
        required: false
        schema:
          type: string
        description: Custom Algolia filter query
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResult'
        '503':
          description: Search service is not configured
...
  /api/search/index/tools:
    post:
      tags:
      - Search
      summary: Index tools
      description: Index all tools in MongoDB to Algolia (asynchronous operation)
      operationId: indexTools
      parameters:
      - name: batch_size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 1000
          default: 100
        description: Number of tools to index in each batch
      responses:
        '202':
          description: Indexing started
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: processing
                  message:
                    type: string
                    example: Indexing tools to Algolia in the background
        '503':
          description: Search service is not configured
...
  /api/search/index/tool/{tool_id}:
    post:
      tags:
      - Search
      summary: Index single tool
      description: Index a single tool to Algolia
      operationId: indexSingleTool
      parameters:
      - name: tool_id
        in: path
        required: true
        schema:
          type: string
        description: ID of the tool to index
      responses:
        '200':
          description: Tool indexed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Tool indexed successfully
                  object_id:
                    type: string
        '404':
          description: Tool not found
        '503':
          description: Search service is not configured
    delete:
      tags:
      - Search
      summary: Delete tool from index
      description: Delete a tool from the Algolia index
      operationId: deleteToolFromIndex
      parameters:
      - name: tool_id
        in: path
        required: true
        schema:
          type: string
        description: ID of the tool to delete from the index
      responses:
        '200':
          description: Tool deleted from index successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Tool deleted from index successfully
        '404':
          description: Tool not found in index
        '503':
          description: Search service is not configured
...
  /tools:
    get:
      tags:
      - Tools
      summary: List tools
      description: List all tools with pagination
      operationId: listTools
      parameters:
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
        description: Number of items to skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 1000
          default: 100
        description: Maximum number of items to return
      responses:
        '200':
          description: List of tools with pagination info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedToolsResponse'
        '401':
          description: Unauthorized
    post:
      tags:
      - Tools
      summary: Create tool
      description: Create a new tool
      operationId: createNewTool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolCreate'
      responses:
        '201':
          description: Tool created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolResponse'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /tools/search:
    get:
      tags:
      - Tools
      summary: Search tools
      description: Search for tools by name or description
      operationId: searchToolsEndpoint
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
        description: Search query
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
        description: Number of items to skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 1000
          default: 100
        description: Maximum number of items to return
      responses:
        '200':
          description: Search results with pagination info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedToolsResponse'
        '401':
          description: Unauthorized
  /tools/{tool_id}:
    get:
      tags:
      - Tools
      summary: Get tool
      description: Get a specific tool by its UUID
      operationId: getTool
      parameters:
      - name: tool_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Tool UUID
      responses:
        '200':
          description: Tool details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolResponse'
        '401':
          description: Unauthorized
        '404':
          description: Tool not found
    put:
      tags:
      - Tools
      summary: Update tool
      description: Update an existing tool
      operationId: updateExistingTool
      parameters:
      - name: tool_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Tool UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolUpdate'
      responses:
        '200':
          description: Tool updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolResponse'
        '401':
          description: Unauthorized
        '404':
          description: Tool not found
    delete:
      tags:
      - Tools
      summary: Delete tool
      description: Delete a tool
      operationId: deleteExistingTool
      parameters:
      - name: tool_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Tool UUID
      responses:
        '204':
          description: Tool deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Tool not found
  /tools/unique/{unique_id}:
    get:
      tags:
      - Tools
      summary: Get tool by unique ID
      description: Get a specific tool by its unique_id
      operationId: getToolByUniqueIdentifier
      parameters:
      - name: unique_id
        in: path
        required: true
        schema:
          type: string
        description: Tool unique ID
      responses:
        '200':
          description: Tool details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolResponse'
        '401':
          description: Unauthorized
        '404':
          description: Tool not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SearchResult:
      type: object
      properties:
        hits:
          type: array
          items:
            type: object
            additionalProperties: true
        page:
          type: integer
        per_page:
          type: integer
        total_hits:
          type: integer
        total_pages:
          type: integer
        processing_time_ms:
          type: integer
        query:
          type: string
...
    ToolCreate:
      type: object
      required:
      - name
      - description
      - unique_id
      - link
      - price
      properties:
        name:
          type: string
          description: Tool name
        description:
          type: string
          description: Tool description
        unique_id:
          type: string
          description: Unique identifier for the tool
        link:
          type: string
          format: uri
          description: Tool website URL
        price:
          type: string
          description: Pricing information
        rating:
          type: string
          description: Tool rating
        category:
          type: string
          nullable: true
          description: Primary category for the tool
        features:
          type: array
          items:
            type: string
          description: Features of the tool
        is_featured:
          type: boolean
          description: Whether the tool is featured
        category_ids:
          type: array
          items:
            type: string
          description: Categories this tool belongs to
        pricing_type:
          type: string
          enum:
          - free
          - freemium
          - paid
          - subscription
          description: Pricing type
        website_url:
          type: string
          format: uri
          description: Tool website URL
        api_url:
          type: string
          format: uri
          description: Tool API URL
        documentation_url:
          type: string
          format: uri
          description: Tool documentation URL
        logo_url:
          type: string
          format: uri
          description: Tool logo URL
        tags:
          type: array
          items:
            type: string
          description: Tags associated with this tool
        metadata:
          type: object
          additionalProperties: true
          description: Additional metadata
    ToolUpdate:
      type: object
      properties:
        name:
          type: string
          description: Tool name
        description:
          type: string
          description: Tool description
        unique_id:
          type: string
          description: Unique identifier for the tool
        link:
          type: string
          format: uri
          description: Tool website URL
        price:
          type: string
          description: Pricing information
        rating:
          type: string
          description: Tool rating
        category:
          type: string
          nullable: true
          description: Primary category for the tool
        features:
          type: array
          items:
            type: string
          description: Features of the tool
        is_featured:
          type: boolean
          description: Whether the tool is featured
        category_ids:
          type: array
          items:
            type: string
          description: Categories this tool belongs to
        pricing_type:
          type: string
          enum:
          - free
          - freemium
          - paid
          - subscription
          description: Pricing type
        website_url:
          type: string
          format: uri
          description: Tool website URL
        api_url:
          type: string
          format: uri
          description: Tool API URL
        documentation_url:
          type: string
          format: uri
          description: Tool documentation URL
        logo_url:
          type: string
          format: uri
          description: Tool logo URL
        tags:
          type: array
          items:
            type: string
          description: Tags associated with this tool
        metadata:
          type: object
          additionalProperties: true
          description: Additional metadata
    ToolResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Tool ID
        name:
          type: string
          description: Tool name
        description:
          type: string
          description: Tool description
        unique_id:
          type: string
          description: Unique identifier for the tool
        link:
          type: string
          format: uri
          description: Tool website URL
        price:
          type: string
          description: Pricing information
        rating:
          type: string
          description: Tool rating
        saved_numbers:
          type: integer
          nullable: true
          description: Number of users who saved this tool
        category:
          type: string
          nullable: true
          description: Primary category for the tool
        features:
          type: array
          items:
            type: string
          description: Features of the tool
        is_featured:
          type: boolean
          description: Whether the tool is featured
        saved_by_user:
          type: boolean
          description: Whether the authenticated user has saved this tool
        created_at:
          type: string
          format: date-time
          description: Tool creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Tool last update timestamp
        metadata:
          type: object
          additionalProperties: true
          description: Additional metadata
    PaginatedToolsResponse:
      type: object
      properties:
        tools:
          type: array
          items:
            $ref: '#/components/schemas/ToolResponse'
          description: List of tools
        total:
          type: integer
          description: Total number of tools
        skip:
          type: integer
          description: Number of items skipped
        limit:
          type: integer
          description: Maximum number of items returned
...
security:
- bearerAuth: []